Save prompts in constants for easier editing
